% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{genotypeFasta}
\alias{genotypeFasta}
\title{Return the nucleotide sequences of a genotype}
\usage{
genotypeFasta(
  genotype,
  germline_db,
  novel = NA,
  include_unseen = FALSE,
  strip_d = TRUE
)
}
\arguments{
\item{genotype}{\code{data.frame} of alleles denoting a genotype,
as returned by \link{inferGenotype}.}

\item{germline_db}{vector of named nucleotide germline sequences
matching the alleles detailed in \code{genotype}.}

\item{novel}{an optional \code{data.frame} containing putative
novel alleles of the type returned by
\link{findNovelAlleles}.}

\item{include_unseen}{if \code{TRUE}, include germline database alleles for
genes that are not present in \code{genotype}. For
genes present in \code{genotype}, include only the
genotyped alleles.}

\item{strip_d}{if \code{TRUE} (default) remove the "D" from the end of
gene annotations (denoting a duplicate gene in the locus)
when matching genotype alleles to \code{germline_db}. If
\code{FALSE}, alleles are matched exactly.}
}
\value{
A named vector of strings containing the germline nucleotide
          sequences of the alleles in the provided genotype.
}
\description{
\code{genotypeFasta} converts a genotype table into a vector of nucleotide
sequences.
}
\examples{
# Find the sequences that correspond to the genotype
genotype_db <- genotypeFasta(SampleGenotype, SampleGermlineIGHV, SampleNovel)

}
\seealso{
\link{inferGenotype}
}
