% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{novel_df}
\alias{novel_df}
\title{Example of Analyzed Rep-Seq data}
\format{A \code{data.frame} where rows correspond to alleles checked for
polymorphisms and columns give results as well as paramaters used to run
the test.}
\description{
Example VDJ-rearranged immunoglobulin Rep-Seq sequences derived from a single
individual (PGP1), sequenced on the Roche 454 platform, and thought by
IMGT/V-QUEST to utilize IGHV1 family alleles, as processed by
\link{findNovelAlleles}.
}
\references{
Gadala-Maria \emph{et al}. (2015) Automated analysis of
high-throughput B cell sequencing data reveals a high frequency of novel
immunoglobulin V gene segment alleles. \emph{PNAS}. 112(8):E862-70.
}
\keyword{data}
