% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readIgFasta}
\alias{readIgFasta}
\title{Read immunoglobulin sequences}
\usage{
readIgFasta(fasta_file, strip_down_name = TRUE, force_caps = TRUE)
}
\arguments{
\item{fasta_file}{fasta-formatted file of immunoglobulin sequences.}

\item{strip_down_name}{if \code{TRUE}, will extract only the allele name
from the strings fasta file's sequence names.}

\item{force_caps}{if \code{TRUE}, will force nucleotides to
uppercase.}
}
\value{
Named vector of strings representing Ig alleles.
}
\description{
\code{readIgFasta} reads a fasta-formatted file of immunoglobulin (Ig)
sequences and returns a named vector of those sequences.
}
\examples{
\dontrun{
# germlines <- readIgFasta("ighv.fasta")
}

}
\seealso{
\link{writeFasta} to do the inverse.
}
