% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigger.R
\docType{package}
\name{tigger-package}
\alias{tigger-package}
\title{tigger: Infers Novel Immunoglobulin Alleles from Sequencing Data}
\description{
Infers the V genotype of an individual from immunoglobulin (Ig) repertoire sequencing data (AIRR-Seq, Rep-Seq). Includes detection of any novel alleles. This information is then used to correct existing V allele calls from among the sample sequences. Citations: Gadala-Maria, et al (2015) \doi{10.1073/pnas.1417683112}, Gadala-Maria, et al (2019) \doi{10.3389/fimmu.2019.00129}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tigger.readthedocs.io}
  \item Report bugs at \url{https://github.com/immcantation/tigger/issues}
}

}
\author{
\strong{Maintainer}: Susanna Marquez \email{susanna.marquez@yale.edu}

Authors:
\itemize{
  \item Susanna Marquez \email{susanna.marquez@yale.edu}
  \item Daniel Gadala-Maria \email{daniel.gadala-maria@yale.edu}
  \item Moriah Cohen \email{moriah.cohen@biu.ac.il}
  \item Ayelet Peres \email{ayelet.peres@yale.edu}
  \item Jason Vander Heiden \email{jason.vanderheiden@gmail.com}
  \item Gur Yaari \email{gur.yaari@yale.edu}
  \item Steven Kleinstein \email{steven.kleinstein@yale.edu} [copyright holder]
}

}
\keyword{internal}
