% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{writeFasta}
\alias{writeFasta}
\title{Write to a fasta file}
\usage{
writeFasta(named_sequences, file, width = 60, append = FALSE)
}
\arguments{
\item{named_sequences}{vector of named string representing sequences}

\item{file}{the name of the output file.}

\item{width}{the number of characters to be printed per line.
if not between 1 and 255, width with be infinite.}

\item{append}{\code{logical} indicating if the output should be
appended to \code{file} instead of overwriting it}
}
\value{
A named vector of strings representing Ig alleles.
}
\description{
\code{writeFasta} writes a named vector of sequences to a file in fasta
format.
}
\examples{
\dontrun{
# writeFasta(germlines, "ighv.fasta")
}

}
\seealso{
\link{readIgFasta} to do the inverse.
}
