% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cleanSeqs}
\alias{cleanSeqs}
\title{Clean up nucleotide sequences}
\usage{
cleanSeqs(seqs)
}
\arguments{
\item{seqs}{a vector of nucleotide sequences}
}
\value{
A vector of nucleotide sequences
}
\description{
\code{cleanSeqs} capitalizes nucleotides, replaces "." with "-", and then
replaces all characters besides ACGT- with "N".
}
\examples{
# Create messy nucleotide sequences
seqs = c("AGAT.taa-GAG...ATA",
         "GATACAGTXXXXXAGNNNPPPACA")
# Clean them up
cleanSeqs(seqs)

}
\seealso{
\link{sortAlleles} and \link{updateAlleleNames} can
         help format a list of allele names.
}
