% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GermlineIGHV}
\alias{GermlineIGHV}
\title{Human IGHV germlines}
\format{
Values correspond to IMGT-gaped nuceltoide sequences (with
nucleotides capitalized and gaps represented by ".") while names correspond
to stripped-down IMGT allele names (e.g. "IGHV1-18*01").
}
\description{
A \code{character} vector of all human IGHV germline gene segment alleles
in IMGT/GENE-DB (2019-06-01, 372 alleles). 
See IMGT data updates: https://www.imgt.org/IMGTgenedbdoc/dataupdates.html.
}
\references{
\enumerate{
  \item Xochelli, et al. (2014) Immunoglobulin heavy variable (IGHV) genes and 
        alleles: new entities, new names and implications for research and 
        prognostication in chronic lymphocytic leukaemia. Immunogenetics. 67(1):61-6.
}
}
\keyword{data}
