% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{insertPolymorphisms}
\alias{insertPolymorphisms}
\title{Insert polymorphisms into a nucleotide sequence}
\usage{
insertPolymorphisms(sequence, positions, nucleotides)
}
\arguments{
\item{sequence}{starting nucletide sequence.}

\item{positions}{numeric vector of positions which to be changed.}

\item{nucleotides}{character vector of nucletides to which to change the
positions.}
}
\value{
A sequence with the desired nucleotides in the provided locations.
}
\description{
\code{insertPolymorphisms} replaces nucleotides in the desired locations of a
provided sequence.
}
\examples{
insertPolymorphisms("HUGGED", c(1, 6, 2), c("T", "R", "I")) 

}
