% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{insertPolymorphisms}
\alias{insertPolymorphisms}
\title{Insert polymorphisms into a nucleotide sequence}
\usage{
insertPolymorphisms(sequence, positions, nucleotides)
}
\arguments{
\item{sequence}{the starting nucletide sequence}

\item{positions}{a vector of positions which to be changed}

\item{nucleotides}{a vector of nucletides to which to change the
positions}
}
\value{
a sequence with the desired nucleotides in provided locations
}
\description{
\code{insertPolymorphisms} replaces nucleotides in the desired locations of a
provided sequence.
}
\examples{
insertPolymorphisms("hugged", c(1,6,2), c("t","r","i")) 

}
