% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SampleGenotype}
\alias{SampleGenotype}
\title{Example genotype inferrence results}
\format{
A \code{data.frame} where rows correspond to genes carried by an
individual and columns lists the alleles of those genes and their counts.
}
\description{
A \code{data.frame} of genotype inference results from \link{inferGenotype}
after novel allele detection via \link{findNovelAlleles}.
Source data was a collection of V(D)J immunoglobulin sequences derived from a single
individual (PGP1), sequenced on the Roche 454 platform, and assigned by
IMGT/HighV-QUEST to IGHV1 family alleles.
}
\references{
\enumerate{
  \item Gadala-Maria, et al. (2015) Automated analysis of high-throughput B cell 
        sequencing data reveals a high frequency of novel immunoglobulin V gene 
        segment alleles. PNAS. 112(8):E862-70.
}
}
\seealso{
See \link{inferGenotype} for detailed column descriptions.
}
\keyword{data}
