% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SampleGermlineIGHV}
\alias{SampleGermlineIGHV}
\title{Example Human IGHV germlines}
\format{Values correspond to IMGT-gaped nuceltoide sequences (with
nucleotides capitalized and gaps represented by ".") while names correspond
to stripped-down IMGT allele names (e.g. "IGHV1-18*01").}
\description{
A \code{character} vector of all 344 human IGHV germline gene segment alleles
in IMGT/GENE-DB release 201408-4.
}
\references{
\enumerate{
  \item Xochelli, et al. (2014) Immunoglobulin heavy variable (IGHV) genes and 
        alleles: new entities, new names and implications for research and 
        prognostication in chronic lymphocytic leukaemia. Immunogenetics. 67(1):61-6.
}
}
\keyword{data}
