% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SampleNovel}
\alias{SampleNovel}
\title{Example novel allele detection results}
\format{
A \code{data.frame} where rows correspond to alleles checked for
polymorphisms and columns give results as well as paramaters used to run
the test.
}
\description{
A \code{data.frame} of novel allele detection results from \link{findNovelAlleles}. 
Source data was a collection of V(D)J immunoglobulin sequences derived from a single
individual (PGP1), sequenced on the Roche 454 platform, and assigned by
IMGT/HighV-QUEST to IGHV1 family alleles.
}
\references{
\enumerate{
  \item Gadala-Maria, et al. (2015) Automated analysis of high-throughput B cell 
        sequencing data reveals a high frequency of novel immunoglobulin V gene 
        segment alleles. PNAS. 112(8):E862-70.
}
}
\seealso{
See \link{findNovelAlleles} for detailed column descriptions.
}
\keyword{data}
